Add claim mapping
The name default no longer makes sense.  I have discovered a document[1]
which seems to have indexed the various services and how to configure a
statement for them in SNS.  The claims are now based on that mapping,
which is many-to-one.

This allows the default mapping to be unused in some cases.

1: https://github.com/widdix/aws-cf-templates/blob/master/operations/alert.yaml